add resource loaded test
